added RectangularWaveguideTE10 as a subclass of the RectangularWaveguide. this was not as clean as i liked because the internal references to it methods got passed the mode index variables. to get around this error, i made all the methods of teh TE10 mode take *args, just to absorb the indicies (because i dont care)
i wanted to use properties for TE10 but i would have had to re-write the functions,
so that it wouldnt have passed any arguments. i chose not to do this because id like 
to have all teh physics written only once, in case i want to update it latter, or adjust it. 


now i am going to do a bunch a re-nameing of transmissionLine so that its consistent with
the convections of my other files. i am doing this now so because i am going to import 
RectangularWaveguideTE10 as a transmission line class, which is going to inherit the 
pertinent methods which our createNetwork function is going to call. 

i might mess this re-nameing up so i am committing now.