two port cal appears to work although new shit has come to light. i dont know if the s-parameters of the error networks can be uniquely determined.
it seems most methods only solve it enough so that its effects
can be removed. this is not a problem in itself, but i will have
to change some properties of Calibration type, like what 
would Calibration.error_network mean? 

also to do a two-port cal, by using t-parameters, the s21's on 
any reflect standards must be perturbed from zero to avoid the 
1/s21, inf problem. im not sure how others handle this.