made BIG CHANGES to Calibration class. and BROKE parametersized one port calibration type.
this was done so that a calibration can be initialiazed with as 
little input from user as possible. so the cal type is infered from 
number of ports of one of the measured networks, as well as 
the frequency information.

also i implemented some basic checking on consistency of  calibration
data, like same # of measurements as ideals, frequency info is identical
for all measured and ideals, etc.