!!MAJOR CHANGES to transmission line functions !!
for clarity and consistency i am defining for a tline

S21 = exp(-theta) = exp(-gamma(f)*d) 

this implies:
	+real(gamma(f)) = attenuation
	+imag(gamma(f)) = forward propagating
and 
	gamma = sqrt(Z'*Y')
	Z0 = sqrt(Z'/Y')

where Z' and Y' are distributed impedance and admittance, respectively. 


i have changed the k_z function in rectangular waveguide to reflect this, 
but have yet to test any of it.